Clean up local workbench helper source
